+++ src/Tests/YamlMap.Tests/Serialization/Mappers/MapperFactoryTests.cs
@@ -1,8 +1,5 @@
-﻿using System;
-using System.Collections;
+﻿using System.Collections;
 using System.Collections.Generic;
-using System.Text;
-using NUnit.Framework;
 using YamlMap.Serialization.Mappers;
 using ArrayMapper = YamlMap.Serialization.Mappers.ArrayMapper;
 
@@ -12,39 +9,48 @@ namespace YamlMap.Tests.Serialization.Mappers
     public class MapperFactoryTests
     {
         [Test]
-        public void YamlMap_Serialization_MapperFactory_List()
+        public void MapperFactory_List()
         {
             var mapper = MapperFactory.GetObjectMapper(new List<MapperItem>(), typeof(MapperItem));
-            Assert.That(mapper, Is.InstanceOf<GenericListMapper>());
+            mapper.Should().BeOfType<GenericListMapper>();
         }
 
         [Test]
-        public void YamlMap_Serialization_MapperFactory_Dictionary()
+        public void MapperFactory_Dictionary()
         {
             var mapper = MapperFactory.GetObjectMapper(new Dictionary<string, string>(), typeof(MapperItem));
-            Assert.That(mapper, Is.InstanceOf<GenericDictionaryMapper>());
+            mapper.Should().BeOfType<GenericDictionaryMapper>();
         }
 
         [Test]
-        public void YamlMap_Serialization_MapperFactory_Array()
+        public void MapperFactory_Array()
         {
             var mapper = MapperFactory.GetObjectMapper(new string[] {}, typeof(MapperItem));
-            Assert.That(mapper, Is.InstanceOf<ArrayMapper>());
+            mapper.Should().BeOfType<ArrayMapper>();
         }
 
         [Test]
-        public void YamlMap_Serialization_MapperFactory_Array2()
+        public void MapperFactory_Array2()
         {
             var mapper = MapperFactory.GetObjectMapper(new ArrayList(), typeof(IList));
-            Assert.That(mapper, Is.InstanceOf<ArrayMapper>());
+            mapper.Should().BeOfType<ArrayMapper>();
+        }
+        
+        [Test]
+        public void MapperFactory_YamlNode()
+        {
+            var mapper = MapperFactory.GetObjectMapper(new YamlNode(), typeof(YamlNode));
+            mapper.Should().BeOfType<YamlNodeMapper>();
         }
 
         [Test]
-        public void YamlMap_Serialization_MapperFactory_Default()
+        public void MapperFactory_Default()
         {
             var mapper = MapperFactory.GetObjectMapper(new MapperItem(), typeof(MapperItem));
             Assert.That(mapper, Is.InstanceOf<DefaultMapper>());
         }
+        
+        
 
         public class MapperItem { }
     }